requirements.txt
torch
ultralytics
torchvision
opencv-python
numpy
pillow
scikit-learn
rawpy
streamlit
matplotlib
seaborn
tqdm
